Download single workbook

+++ scripts/src/plh-copy-app-data.ts
@@ -9,11 +9,13 @@ const APP_DATA_DIR = `../src/data`;
 const APP_PLH_ASSETS_DIR = "../src/assets/plh_assets";
 fs.ensureDirSync(APP_DATA_DIR);
 fs.emptyDirSync(APP_DATA_DIR);
-fs.ensureDirSync(APP_PLH_ASSETS_DIR);
-fs.emptyDirSync(APP_PLH_ASSETS_DIR);
 
 /** A simple script to copy data exported from gdrive and processed for plh into the app data folder */
-function main() {
+export function main(doAssetFolderCheck = true) {
+  if (doAssetFolderCheck) {
+    fs.ensureDirSync(APP_PLH_ASSETS_DIR);
+    fs.emptyDirSync(APP_PLH_ASSETS_DIR);
+  }
   console.log(chalk.yellow("Copying Data To App"));
   const localTsFiles = fs.readdirSync(DATA_INPUT_FOLDER);
   for (const filepath of localTsFiles) {
@@ -25,7 +27,10 @@ function main() {
   fs.copySync(ASSETS_INPUT_FOLDER, APP_PLH_ASSETS_DIR);
   console.log(chalk.green("Data Copied to App"));
 }
-main();
+
+if (process.argv[1] && process.argv[1].indexOf("sync-single") < 0) {
+  main();
+}
 
 /**
  * When copying to the app simply replace the path to local typings imported to
